+++ Coimbra/EventRefHandler.cs
@@ -5,6 +5,5 @@ namespace Coimbra
     /// <summary>
     /// Similar to <see cref="EventHandler{TEventArgs}"/> but using the ref keyword.
     /// </summary>
-    /// <typeparam name="T"></typeparam>
     public delegate void EventRefHandler<T>(object sender, ref T e);
 }