+++ Coimbra/Services/IEventService.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using UnityEngine.Scripting;
 
 namespace Coimbra
@@ -14,14 +15,18 @@ namespace Coimbra
         /// </summary>
         /// <param name="eventCallback">The callback to be added.</param>
         /// <typeparam name="T">The event type.</typeparam>
-        EventHandle AddListener<T>(EventHandler<T> eventCallback);
+        EventHandle AddListener<T>(EventRefHandler<T> eventCallback)
+            where T : IEvent;
 
         /// <summary>
         /// Adds a listener to an event type.
         /// </summary>
         /// <param name="eventCallback">The callback to be added.</param>
+        /// <param name="appendList">List to add the event handle if generated a valid one.</param>
         /// <typeparam name="T">The event type.</typeparam>
-        EventHandle AddListener<T>(EventRefHandler<T> eventCallback);
+        /// <returns>True if a valid <see cref="EventHandle"/> was generated.</returns>
+        bool AddListener<T>(EventRefHandler<T> eventCallback, List<EventHandle> appendList)
+            where T : IEvent;
 
         /// <summary>
         /// Compares the event key for specified type.
@@ -29,7 +34,8 @@ namespace Coimbra
         /// <param name="eventKey">The event key to compare to.</param>
         /// <typeparam name="T">The event type.</typeparam>
         /// <returns>True if the current event key matches the <see cref="eventKey"/> parameter.</returns>
-        bool CompareEventKey<T>(object eventKey);
+        bool CompareEventKey<T>(object eventKey)
+            where T : IEvent;
 
         /// <summary>
         /// Compares the event key for specified type.
@@ -44,7 +50,8 @@ namespace Coimbra
         /// </summary>
         /// <typeparam name="T">The event type.</typeparam>
         /// <returns>True if there is any listeners for this event on this service.</returns>
-        bool HasAnyListeners<T>();
+        bool HasAnyListeners<T>()
+            where T : IEvent;
 
         /// <summary>
         /// Checks if an event type contains any listeners.
@@ -64,11 +71,10 @@ namespace Coimbra
         /// Invokes the specified event type for all its listeners with a default constructed data.
         /// </summary>
         /// <param name="eventSender">The object invoking the event.</param>
-        /// <param name="eventType">The event type to be constructed and sent.</param>
         /// <param name="eventKey">The encapsulation key for the event.</param>
-        /// <param name="ignoreException">If true, the method will fail silently if <see cref="eventKey"/> doesn't match the current event key.</param>
-        /// <exception cref="InvalidOperationException">If <see cref="ignoreException"/> is false and <see cref="eventKey"/> doesn't match the current event key.</exception>
-        void Invoke(object eventSender, Type eventType, object eventKey = null, bool ignoreException = false);
+        /// <typeparam name="T">The event type to be constructed and sent.</typeparam>
+        void Invoke<T>(object eventSender, object eventKey = null)
+            where T : IEvent, new();
 
         /// <summary>
         /// Invokes the specified event type for all its listeners.
@@ -76,10 +82,9 @@ namespace Coimbra
         /// <param name="eventSender">The object invoking the event.</param>
         /// <param name="eventData">The event data to be sent.</param>
         /// <param name="eventKey">The encapsulation key for the event.</param>
-        /// <param name="ignoreException">If true, the method will fail silently if <see cref="eventKey"/> doesn't match the current event key.</param>
         /// <typeparam name="T">The event type.</typeparam>
-        /// <exception cref="InvalidOperationException">If <see cref="ignoreException"/> is false and <see cref="eventKey"/> doesn't match the current event key.</exception>
-        void Invoke<T>(object eventSender, T eventData, object eventKey = null, bool ignoreException = false);
+        void Invoke<T>(object eventSender, T eventData, object eventKey = null)
+            where T : IEvent;
 
         /// <summary>
         /// Invokes the specified event type for all its listeners.
@@ -87,36 +92,30 @@ namespace Coimbra
         /// <param name="eventSender">The object invoking the event.</param>
         /// <param name="eventData">The event data to be sent.</param>
         /// <param name="eventKey">The encapsulation key for the event.</param>
-        /// <param name="ignoreException">If true, the method will fail silently if <see cref="eventKey"/> doesn't match the current event key.</param>
         /// <typeparam name="T">The event type.</typeparam>
-        /// <exception cref="InvalidOperationException">If <see cref="ignoreException"/> is false and <see cref="eventKey"/> doesn't match the current event key.</exception>
-        void Invoke<T>(object eventSender, ref T eventData, object eventKey = null, bool ignoreException = false);
+        void Invoke<T>(object eventSender, ref T eventData, object eventKey = null)
+            where T : IEvent;
 
         /// <summary>
         /// Removes all listeners from all event types.
         /// </summary>
         /// <param name="serviceKey">The encapsulation key for the service.</param>
-        /// <param name="ignoreException">If true, the method will fail silently if <see cref="serviceKey"/> doesn't match the current service key.</param>
-        /// <exception cref="InvalidOperationException">If <see cref="ignoreException"/> is false and <see cref="serviceKey"/> doesn't match the current service key.</exception>
-        void RemoveAllListeners(object serviceKey = null, bool ignoreException = false);
+        void RemoveAllListeners(object serviceKey = null);
 
         /// <summary>
         /// Removes all listeners from an event type.
         /// </summary>
         /// <param name="eventKey">The encapsulation key for the event.</param>
-        /// <param name="ignoreException">If true, the method will fail silently if <see cref="eventKey"/> doesn't match the current event key.</param>
         /// <typeparam name="T">The event type.</typeparam>
-        /// <exception cref="InvalidOperationException">If <see cref="ignoreException"/> is false and <see cref="eventKey"/> doesn't match the current event key.</exception>
-        void RemoveAllListeners<T>(object eventKey = null, bool ignoreException = false);
+        void RemoveAllListeners<T>(object eventKey = null)
+            where T : IEvent;
 
         /// <summary>
         /// Removes all listeners from an event type.
         /// </summary>
         /// <param name="eventType">The event type.</param>
         /// <param name="eventKey">The encapsulation key for the event.</param>
-        /// <param name="ignoreException">If true, the method will fail silently if <see cref="eventKey"/> doesn't match the current event key.</param>
-        /// <exception cref="InvalidOperationException">If <see cref="ignoreException"/> is false and <see cref="eventKey"/> doesn't match the current event key.</exception>
-        void RemoveAllListeners(Type eventType, object eventKey = null, bool ignoreException = false);
+        void RemoveAllListeners(Type eventType, object eventKey = null);
 
         /// <summary>
         /// Removes a listener from an event with its handle.
@@ -129,7 +128,8 @@ namespace Coimbra
         /// </summary>
         /// <param name="eventKey">The encapsulation key for the event.</param>
         /// <typeparam name="T">The event type.</typeparam>
-        void ResetEventKey<T>(object eventKey);
+        void ResetEventKey<T>(object eventKey)
+            where T : IEvent;
 
         /// <summary>
         /// Resets the encapsulation key for an event type;
@@ -143,6 +143,7 @@ namespace Coimbra
         /// </summary>
         /// <param name="eventKey">The encapsulation key for the event.</param>
         /// <typeparam name="T">The event type.</typeparam>
-        void SetEventKey<T>(object eventKey);
+        void SetEventKey<T>(object eventKey)
+            where T : IEvent;
     }
 }