+++ Coimbra/Systems/EventSystem.cs
@@ -1,6 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Diagnostics;
+using System.Runtime.CompilerServices;
 using UnityEngine;
+using Debug = UnityEngine.Debug;
 
 namespace Coimbra
 {
@@ -35,6 +38,7 @@ namespace Coimbra
 
         private delegate bool RemoveHandler(EventHandle key);
 
+        private const string InvalidEventKeyMessageFormat = "The event key \"{0}\" doesn't match the current set key \"{1}\" for type \"{2}\"";
         private readonly object _serviceKey;
         private readonly Dictionary<Type, Event> _events = new Dictionary<Type, Event>();
 
@@ -47,22 +51,11 @@ namespace Coimbra
         public ServiceLocator OwningLocator { get; set; }
 
         /// <inheritdoc/>
-        public EventHandle AddListener<T>(EventHandler<T> eventCallback)
-        {
-            if (eventCallback == null)
-            {
-                return new EventHandle();
-            }
-
-            return AddListener(delegate(object sender, ref T e)
-            {
-                eventCallback.Invoke(sender, e);
-            });
-        }
-
-        /// <inheritdoc/>
         public EventHandle AddListener<T>(EventRefHandler<T> eventCallback)
+            where T : IEvent
         {
+            CheckType(typeof(T));
+
             if (eventCallback == null)
             {
                 return new EventHandle();
@@ -90,7 +83,24 @@ namespace Coimbra
         }
 
         /// <inheritdoc/>
+        public bool AddListener<T>(EventRefHandler<T> eventCallback, List<EventHandle> appendList)
+            where T : IEvent
+        {
+            EventHandle eventHandle = AddListener(eventCallback);
+
+            if (!eventHandle.IsValid)
+            {
+                return false;
+            }
+
+            appendList?.Add(eventHandle);
+
+            return true;
+        }
+
+        /// <inheritdoc/>
         public bool CompareEventKey<T>(object eventKey)
+            where T : IEvent
         {
             return CompareEventKey(typeof(T), eventKey);
         }
@@ -98,6 +108,8 @@ namespace Coimbra
         /// <inheritdoc/>
         public bool CompareEventKey(Type eventType, object eventKey)
         {
+            CheckType(eventType);
+
             return _events.TryGetValue(eventType, out Event e) ? e.Key == eventKey : eventKey == null;
         }
 
@@ -110,6 +122,7 @@ namespace Coimbra
 
         /// <inheritdoc/>
         public bool HasAnyListeners<T>()
+            where T : IEvent
         {
             return HasAnyListeners(typeof(T));
         }
@@ -117,6 +130,8 @@ namespace Coimbra
         /// <inheritdoc/>
         public bool HasAnyListeners(Type eventType)
         {
+            CheckType(eventType);
+
             return _events.TryGetValue(eventType, out Event e) && e.Handles.Count > 0;
         }
 
@@ -127,20 +142,26 @@ namespace Coimbra
         }
 
         /// <inheritdoc/>
-        public void Invoke(object eventSender, Type eventType, object eventKey = null, bool ignoreException = false)
+        public void Invoke<T>(object eventSender, object eventKey = null)
+            where T : IEvent, new()
         {
-            Invoke(eventSender, Activator.CreateInstance(eventType), eventKey, ignoreException);
+            T eventData = new T();
+            Invoke(eventSender, ref eventData, eventKey);
         }
 
         /// <inheritdoc/>
-        public void Invoke<T>(object eventSender, T eventData, object eventKey = null, bool ignoreException = false)
+        public void Invoke<T>(object eventSender, T eventData, object eventKey = null)
+            where T : IEvent
         {
-            Invoke(eventSender, ref eventData, eventKey, ignoreException);
+            Invoke(eventSender, ref eventData, eventKey);
         }
 
         /// <inheritdoc/>
-        public void Invoke<T>(object eventSender, ref T eventData, object eventKey = null, bool ignoreException = false)
+        public void Invoke<T>(object eventSender, ref T eventData, object eventKey = null)
+            where T : IEvent
         {
+            CheckType(typeof(T));
+
             if (!_events.TryGetValue(typeof(T), out Event e))
             {
                 return;
@@ -148,12 +169,16 @@ namespace Coimbra
 
             if (e.Key != null && e.Key != eventKey)
             {
-                if (ignoreException)
-                {
-                    return;
-                }
+                Debug.LogErrorFormat(InvalidEventKeyMessageFormat, eventKey, e.Key, typeof(T));
 
-                throw new InvalidOperationException();
+                return;
+            }
+
+            if (e.IsInvoking)
+            {
+                Debug.LogError($"{typeof(T)} is already being invoked! Skipping its invocation to avoid a stack overflow.");
+
+                return;
             }
 
             e.IsInvoking = true;
@@ -185,16 +210,13 @@ namespace Coimbra
         }
 
         /// <inheritdoc/>
-        public void RemoveAllListeners(object serviceKey = null, bool ignoreException = false)
+        public void RemoveAllListeners(object serviceKey = null)
         {
             if (_serviceKey != null && _serviceKey != serviceKey)
             {
-                if (ignoreException)
-                {
-                    return;
-                }
+                Debug.LogError($"The service key \"{serviceKey}\" doesn't match the current set key \"{_serviceKey}\"");
 
-                throw new InvalidOperationException();
+                return;
             }
 
             foreach (Event e in _events.Values)
@@ -216,14 +238,17 @@ namespace Coimbra
         }
 
         /// <inheritdoc/>
-        public void RemoveAllListeners<T>(object eventKey = null, bool ignoreException = false)
+        public void RemoveAllListeners<T>(object eventKey = null)
+            where T : IEvent
         {
-            RemoveAllListeners(typeof(T), eventKey, ignoreException);
+            RemoveAllListeners(typeof(T), eventKey);
         }
 
         /// <inheritdoc/>
-        public void RemoveAllListeners(Type eventType, object eventKey = null, bool ignoreException = false)
+        public void RemoveAllListeners(Type eventType, object eventKey = null)
         {
+            CheckType(eventType);
+
             if (!_events.TryGetValue(eventType, out Event e))
             {
                 return;
@@ -231,12 +256,9 @@ namespace Coimbra
 
             if (e.Key != null && e.Key != eventKey)
             {
-                if (ignoreException)
-                {
-                    return;
-                }
+                Debug.LogErrorFormat(InvalidEventKeyMessageFormat, eventKey, e.Key, eventType);
 
-                throw new InvalidOperationException();
+                return;
             }
 
             if (e.IsInvoking)
@@ -274,6 +296,7 @@ namespace Coimbra
 
         /// <inheritdoc/>
         public void ResetEventKey<T>(object eventKey)
+            where T : IEvent
         {
             ResetEventKey(typeof(T), eventKey);
         }
@@ -281,6 +304,8 @@ namespace Coimbra
         /// <inheritdoc/>
         public void ResetEventKey(Type eventType, object eventKey)
         {
+            CheckType(eventType);
+
             if (eventKey == null || !_events.TryGetValue(eventType, out Event e))
             {
                 return;
@@ -288,7 +313,9 @@ namespace Coimbra
 
             if (e.Key != null && e.Key != eventKey)
             {
-                throw new InvalidOperationException();
+                Debug.LogErrorFormat(InvalidEventKeyMessageFormat, eventKey, e.Key, eventType);
+
+                return;
             }
 
             e.Key = null;
@@ -296,12 +323,17 @@ namespace Coimbra
 
         /// <inheritdoc/>
         public void SetEventKey<T>(object eventKey)
+            where T : IEvent
         {
+            CheckType(typeof(T));
+
             if (_events.TryGetValue(typeof(T), out Event e))
             {
                 if (e.Key != null && e.Key != eventKey)
                 {
-                    throw new InvalidOperationException();
+                    Debug.LogErrorFormat(InvalidEventKeyMessageFormat, eventKey, e.Key, typeof(T));
+
+                    return;
                 }
 
                 e.Key = eventKey;
@@ -320,7 +352,22 @@ namespace Coimbra
         [RuntimeInitializeOnLoadMethod(RuntimeInitializeLoadType.SubsystemRegistration)]
         private static void Initialize()
         {
-            ServiceLocator.SetDefaultCreateCallback(Create, false);
+            ServiceLocator.Shared.SetCreateCallback(Create, false);
+        }
+
+        [Conditional("UNITY_EDITOR")]
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        private static void CheckType(Type type, [CallerMemberName] string memberName = null)
+        {
+            if (type.IsInterface)
+            {
+                throw new ArgumentOutOfRangeException($"\"{nameof(IEventService)}.{memberName}\" requires a non-interface type argument!");
+            }
+
+            if (!typeof(IEvent).IsAssignableFrom(type))
+            {
+                throw new ArgumentOutOfRangeException($"\"{nameof(IEventService)}.{memberName}\" requires a type that implements \"{typeof(IEvent)}\"!");
+            }
         }
     }
 }