+++ Coimbra/Systems/TimerSystem.cs
@@ -147,7 +147,7 @@ namespace Coimbra
         [RuntimeInitializeOnLoadMethod(RuntimeInitializeLoadType.SubsystemRegistration)]
         private static void Initialize()
         {
-            ServiceLocator.SetDefaultCreateCallback(Create, false);
+            ServiceLocator.Shared.SetCreateCallback(Create, false);
         }
     }
 }