+++ Coimbra/Utilities/ObjectUtility.cs
@@ -9,7 +9,7 @@ namespace Coimbra
     public static class ObjectUtility
     {
         /// <summary>
-        /// Get a valid object to be used with ?. and ?? operators.
+        /// Gets a valid object to be used with ?. and ?? operators.
         /// </summary>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static T GetValid<T>(this T o)
@@ -35,6 +35,18 @@ namespace Coimbra
             }
 
             return o != null;
+        }
+
+        /// <summary>
+        /// Safe way to check if an object is valid even if the object is an Unity <see cref="Object"/> and got destroyed already, getting a valid object to be used with ?. and ?? operators too.
+        /// </summary>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static bool TryGetValid<T>(this T o, out T valid)
+            where T : class
+        {
+            valid = GetValid(o);
+
+            return valid != null;
         }
     }
 }